requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.2
Django==3.1.5
django-environ==0.4.5
django-modeltranslation==0.17.3
django-resized==0.3.11
filelock==3.0.12
importlib-metadata==4.6.1
numpy==1.21.0
opencv-python==4.5.3.56
pbr==5.6.0
Pillow==8.3.0
pytz==2020.5
six==1.16.0
sqlparse==0.4.1
stevedore==3.3.0
typing-extensions==3.10.0.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
zipp==3.5.0
